+++ MOP/src/Misc/MopSettings.cs
@@ -11,11 +11,13 @@
         public static bool ToggleVehicles { get; set; }
         public static bool ToggleItems { get; set; }
 
+        public static int TrafficLimit { get; set; }
+
         public static bool EnableObjectOcclusion { get; set; }
         public static int OcclusionSamples = 120;
         public static int ViewDistance = 400;
         public static int OcclusionSampleDelay = 1;
-        public static int OcclusionHideDelay = 3;
+        public static int OcclusionHideDelay = -1;
         public static int MinOcclusionDistance = 50;
         public static int OcclusionMethod = 1;
 
@@ -28,11 +30,13 @@
             ToggleVehicles = (bool)MOP.toggleVehicles.GetValue();
             ToggleItems = (bool)MOP.toggleItems.GetValue();
 
+            TrafficLimit = GetVehicleLimit();
+            MSCLoader.ModConsole.Print(TrafficLimit);
+
             EnableObjectOcclusion = (bool)MOP.enableObjectOcclusion.GetValue();
-            OcclusionSamples = int.Parse(MOP.occlusionSamples.GetValue().ToString());
+            OcclusionSamples = GetOcclusionSamples();
             ViewDistance = int.Parse(MOP.occlusionDistance.GetValue().ToString());
             OcclusionSampleDelay = int.Parse(MOP.occlusionSampleDelay.GetValue().ToString());
-            OcclusionHideDelay = int.Parse(MOP.occlusionHideDelay.GetValue().ToString());
             MinOcclusionDistance = int.Parse(MOP.minOcclusionDistance.GetValue().ToString());
 
             OcclusionMethod = GetOcclusionMethod();
@@ -65,6 +69,43 @@
                 return 2;
 
             return 1;
+        }
+
+        static int GetOcclusionSamples()
+        {
+            if ((bool)MOP.occlusionSamplesLowest.GetValue())
+                return 10;
+
+            if ((bool)MOP.occlusionSamplesLower.GetValue())
+                return 30;
+
+            if ((bool)MOP.occlusionSamplesLow.GetValue())
+                return 60;
+
+            if ((bool)MOP.occlusionSamplesDetailed.GetValue())
+                return 120;
+
+            if ((bool)MOP.occlusionSamplesVeryDetailed.GetValue())
+                return 240;
+
+            return 120;
+        }
+
+        static int GetVehicleLimit()
+        {
+            if ((bool)MOP.highwayTrafficDensityAll.GetValue())
+                return -1;
+
+            if ((bool)MOP.highwayTrafficDensityMost.GetValue())
+                return 7;
+
+            if ((bool)MOP.highwayTrafficDensityHalf.GetValue())
+                return 5;
+
+            if ((bool)MOP.highwayTrafficDensityQuarter.GetValue())
+                return 3;
+
+            return -1;
         }
     }
 }